+++ apps/frontend-personal-space/src/app/pages/Personal-billing.tsx
@@ -1,7 +1,7 @@
 import { useNavigate, useParams } from 'react-router-dom';
 import { Box, Button, Typography } from '@mui/material';
 import { SSEProvider, useSSE } from 'react-hooks-sse';
-import { TableBillingShell } from '@spos/ui/common';
+import { TableBillingShell, GroupClosedListener } from '@spos/ui/common';
 import { useMutation } from '@tanstack/react-query';
 import {
   Configuration,
@@ -203,7 +203,7 @@ export function PersonalBilling() {
     },
     onSuccess: (data) => {
       if (data.data) {
-        navigate('/'+groupId+'/'+tableNumber+"/"+ownerId);
+        navigate('/thanks');
       }
     }
   });
@@ -227,12 +227,14 @@ export function PersonalBilling() {
 
       <SSEProvider
         endpoint={
-          'http://localhost:3002/api/payments/sse/customer-items/' +
+          `${
+      import.meta.env.VITE_PAYMENT_SHARING_BASE_URL}/api/payments/sse/customer-items/` +
           groupId +
           '/' +
           ownerId
         }
       >
+        <GroupClosedListener />
         <Box
           sx={{
             height: '85dvh',