+++ apps/payment-sharing/src/app/payment/payment.controller.ts
@@ -1,25 +1,22 @@
 import { Body, Controller, Get, Param, Post, Sse } from '@nestjs/common';
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { PaymentService } from './payment.service';
-import {
-  BehaviorSubject,
-  filter,
-  fromEvent,
-  merge,
-  Observable,
-  switchMap,
-} from 'rxjs';
+import { Observable } from 'rxjs';
 import { ItemRequestDto } from './Item.dto';
 import { PaymentResponseTableDTO } from './payment-response.dto';
 import { SelectedByCustomerDTO } from './selected-by-customer.dto';
-import { context, propagation, trace } from '@opentelemetry/api';
+import { trace } from '@opentelemetry/api';
+import { SseService } from './sse.service';
 
 @Controller('payments')
 @ApiTags('payments')
 export class PaymentController {
   private readonly sse = trace.getTracer('SSE');
 
-  constructor(private readonly paymentService: PaymentService) {}
+  constructor(
+    private readonly paymentService: PaymentService,
+    private readonly sseService: SseService
+  ) {}
 
   @Post('take-item')
   //@ApiOperation({ summary: 'Take an item from the center table' })
@@ -74,48 +71,7 @@ export class PaymentController {
     @Param('group_id') group_id: string,
     @Param('owner_id') owner_id: string
   ) {
-    const obs = fromEvent(
-      this.paymentService.eventEmitter,
-      'update-payment'
-    ).pipe(
-      filter(
-        (data: { group_id: string; owner_id: string; traceContext: any }) =>
-          data.group_id === group_id && data.owner_id === owner_id
-      ),
-      switchMap(async (data) => {
-        const extractedContext = propagation.extract(
-          context.active(),
-          data.traceContext
-        ); // Extract the OpenTelemetry context
-        return await context.with(extractedContext, async () => {
-          return await this.sse.startActiveSpan(
-            'sseCustomerItems',
-            async (span) => {
-              span.addEvent('sseCustomerItems');
-              span.setAttribute('group_id', group_id);
-              span.setAttribute('owner_id', owner_id);
-              return {
-                data: await this.paymentService.getCustomerItems(
-                  group_id,
-                  owner_id
-                ),
-              };
-            }
-          );
-        });
-      })
-    );
-    // make a separate observable to send the initial data
-    const initialData = await this.paymentService.getCustomerItems(
-      group_id,
-      owner_id
-    );
-    const sub = new BehaviorSubject<{
-      data: SelectedByCustomerDTO[];
-    }>({
-      data: initialData,
-    });
-    return merge(obs, sub).pipe();
+    return this.sseService.getCustomerItemsSee(group_id, owner_id);
   }
 
   @Sse('sse/table-items/:group_id')
@@ -128,48 +84,15 @@ export class PaymentController {
   })
   async sseTableItems(
     @Param('group_id') group_id: string
-  ): Promise<Observable<{ data: PaymentResponseTableDTO[] }>> {
-    const observalbe = fromEvent(
-      this.paymentService.eventEmitter,
-      'update-payment'
-    ).pipe(
-      filter(
-        (data: { group_id: string; traceContext: any }) =>
-          data.group_id === group_id
-      ),
-      switchMap(async (data) => {
-        const extractedContext = propagation.extract(
-          context.active(),
-          data.traceContext
-        ); // Extract the OpenTelemetry context
-        return await context.with(extractedContext, async () => {
-          return await this.sse.startActiveSpan(
-            'sseTableItems',
-            async (span) => {
-              span.addEvent('sseTableItems');
-              span.setAttribute('group_id', group_id);
-              const items = await this.paymentService.getGroupItems(group_id);
-              return { data: items };
-            }
-          );
-        });
-      })
-    );
-    // make a separate observable to send the initial data
-    const initialData = await this.paymentService.getGroupItems(group_id);
-    const sub = new BehaviorSubject<{
-      data: PaymentResponseTableDTO[];
-    }>({
-      data: initialData,
-    });
-    return merge(observalbe, sub).pipe();
+  ) {
+    return this.sseService.getTableItemsSee(group_id);
   }
 
   @Post('pay/:groupId/:ownerId')
   @ApiResponse({
     status: 201,
     description: 'Payment made successfully',
-    type: Boolean
+    type: Boolean,
   })
   async makePayment(
     @Param('groupId') groupId: string,